Laravel 5.7 compatibility wifth php >= 7.3

+++ database/migrations/2017_10_02_124032_create_menu_item_relation.php
@@ -14,7 +14,7 @@ class CreateMenuItemRelation extends Migration
     public function up()
     {
         Schema::table('novius_menu_items', function (Blueprint $table) {
-            $table->unsignedInteger('menu_id');
+            $table->unsignedBigInteger('menu_id');
         });
     }
